refactor(api): store money/quantities as NUMERIC, map to Decimal
The core tables stored USD values and token quantities as double
precision while the billing tables (invoices, fees, accruals) already
used numeric. Migration 0002 converts the 14 money/quantity/price columns
to numeric (percentages, probabilities and bps stay double precision),
and the FromRow structs that read them now map to rust_decimal::Decimal,
matching the existing billing pattern and the "money is Decimal" rule.

- Serialized Decimal fields carry #[serde(with = "rust_decimal::serde::float")]
  so the JSON wire format stays a number (frontend contract unchanged);
  Cargo.toml gains the serde-with-float feature for the helper.
- f64 still does the float-only statistics (volatility/correlation, prices,
  weights); Decimal is converted to f64 only at those boundaries.
- The trustability view is recreated with sum(total_value_usd)::float8 so
  aum_usd stays an f64 boundary for the tier-cap middleware.

Verified: cargo check --all-targets, clippy -D warnings, 408 tests pass.
Live: dev DB migrated (ledger baseline+0002), portfolio/peg/rebalance/tax/
traction/leaderboard reads return 200 with exact numbers (no strings), and
a peg-rule POST round-trips f64 0.985 -> numeric -> Decimal -> 0.985 number.

+++ apps/api/src/modules/diary/handlers.rs
@@ -6,6 +6,8 @@ use chrono::{DateTime, Utc};
 use serde::Serialize;
 use uuid::Uuid;
 
+use rust_decimal::Decimal;
+
 use crate::error::{AppError, Result};
 use crate::modules::wallet_routes;
 use crate::router::AppState;
@@ -170,7 +172,8 @@ pub struct DecisionLeg {
     pub dest_chain: Option<String>,
     pub src_symbol: Option<String>,
     pub dest_symbol: Option<String>,
-    pub amount_usdc: f64,
+    #[serde(with = "rust_decimal::serde::float")]
+    pub amount_usdc: Decimal,
     pub status: String,
     pub tx_hash: Option<String>,
 }
